+++ crates/core/tedge_agent/src/main.rs
@@ -27,7 +27,6 @@ mod restart_manager;
 mod software_manager;
 mod state_repository;
 mod tedge_operation_converter;
-mod tedge_to_te_converter;
 
 #[derive(Debug, Clone, clap::Parser)]
 #[clap(
