+++ crates/core/tedge_mapper/Cargo.toml
@@ -22,6 +22,7 @@ clock = { workspace = true }
 collectd_ext = { workspace = true }
 flockfile = { workspace = true }
 mqtt_channel = { workspace = true }
+serde_json = { workspace = true }
 tedge_actors = { workspace = true }
 tedge_config = { workspace = true }
 tedge_file_system_ext = { workspace = true }
@@ -30,9 +31,7 @@ tedge_http_ext = { workspace = true }
 tedge_mqtt_ext = { workspace = true }
 tedge_signal_ext = { workspace = true }
 tedge_timer_ext = { workspace = true }
-tedge_utils = { workspace = true, features = [
-    "logging",
-] }
+tedge_utils = { workspace = true, features = ["logging"] }
 tokio = { workspace = true, features = [
     "process",
     "rt",
